Each function passes the correct value or pointer to rules.
The combination of Elem().Interface() does not return a pointer, even if the
original variable was a pointer. Passing by value to validation rules becomes a
problem when the value is a struct with a private mutex. This is the case with
structs generated with protoc-gen-go. `go vet`'s copylocks check identifies the
problem.